feat(0.0.8): Require explicit route symbol creation
Change the behaviour of route symbol creation to require more obvious declaration
of routes.

This change is implemented because I can't prove the stability of imports when
bundling.

This is especially the case when different bundlers bundle different parts of the
program and tree shaking comes into play.

It's not really a step back in the ease of use of the API if you look at the whole
picture. Just more easily copy pasted boilerplate.

This is more valuable for peace of mind.

Maybe in the future we might use string based route keys, though that brings its
own set of problems.

This change breaks the route definition api